[widgets] Move core.Endpoint -> widgets.Endpoint, clean up file structure
