+++ src/main/scala/midas/widgets/Assert.scala
@@ -11,9 +11,9 @@ import scala.collection.immutable.ListMap
 
 import freechips.rocketchip.config.{Parameters}
 import freechips.rocketchip.util.{DecoupledHelper}
-
 import junctions._
-import midas.core.{Endpoint, HostPort}
+
+import midas.core.{HostPort}
 
 
 class AssertBundle(val numAsserts: Int) extends Bundle {
